Resolve Confilct
